Cretae utilities function for mapbox layer
