+++ src/mapbox/layers/Layer.js
@@ -1,4 +1,5 @@
 import { unByKey } from 'ol/Observable';
+import { transformExtent } from 'ol/proj';
 import LayerCommon from '../../common/layers/Layer';
 
 /**
@@ -84,6 +85,25 @@ class Layer extends LayerCommon {
         this.map.off('mousemove', this.onUserMoveCallback);
       }
     }
+  }
+
+  /**
+   * Returns the current extent in mercator coordinates.
+   */
+  getMercatorExtent() {
+    const bounds = this.map.getBounds().toArray();
+    return transformExtent(
+      [...bounds[0], ...bounds[1]],
+      'EPSG:4326',
+      'EPSG:3857',
+    );
+  }
+
+  /**
+   * Returns the equivalent zoom in Openlayers.
+   */
+  getOlZoom() {
+    return this.map.getZoom() + 1;
   }
 
   /**
